Reorder the Transformers
Fix #84
