+++ src/main/scala/io/arlas/data/sql/TransformableDataFrame.scala
@@ -22,6 +22,9 @@ package io.arlas.data.sql
 import io.arlas.data.model.DataModel
 import io.arlas.data.transform._
 import io.arlas.data.transform.ArlasTransformerColumns._
+import io.arlas.data.transform.features.WithArlasTimestamp
+import io.arlas.data.transform.timeseries.WithStateIdFromState
+import io.arlas.data.transform.tools.DataFrameFormatter
 import org.apache.spark.ml.Pipeline
 import org.apache.spark.sql.functions.{col, date_format, lit, struct, to_date}
 import org.apache.spark.sql.types.{DataType, IntegerType, StringType}
@@ -31,7 +34,7 @@ class TransformableDataFrame(df: DataFrame) {
 
   def asArlasFormattedData(dataModel: DataModel,
                            doubleColumns: Vector[String] = Vector()): DataFrame = {
-    df.enrichWithArlas(new DataFrameFormatter(dataModel), new WithArlasTimestamp(dataModel))
+    df.enrichWithArlas(new DataFrameFormatter(dataModel, doubleColumns), new WithArlasTimestamp(dataModel))
       .withColumn(arlasPartitionColumn,
                   date_format(to_date(col(dataModel.timestampColumn), dataModel.timeFormat),
                               "yyyyMMdd").cast(IntegerType))
