+++ src/main/scala/io/arlas/data/transform/timeseries/WithStateIdOnStateChange.scala
@@ -17,25 +17,31 @@
  * under the License.
  */
 
-package io.arlas.data.transform
+package io.arlas.data.transform.timeseries
 
 import io.arlas.data.model.DataModel
+import org.apache.spark.sql.expressions.Window
 import org.apache.spark.sql.functions._
 
 /**
-  * Compute ID column, the same ID is set for all consecutive rows between 2 occurences of fromState, for a same object
+  * Compute ID column, the same ID is set for all consecutive rows with the same state
   * @param dataModel
   * @param stateColumn
-  * @param fromState
   * @param targetIdColumn
   */
-class WithStateIdFromState(dataModel: DataModel,
-                           stateColumn: String,
-                           orderColumn: String,
-                           fromState: String,
-                           targetIdColumn: String)
-    extends WithStateId(dataModel,
-                        stateColumn,
-                        orderColumn,
-                        targetIdColumn,
-                        col(stateColumn).equalTo(fromState))
+class WithStateIdOnStateChange(dataModel: DataModel,
+                               stateColumn: String,
+                               orderColumn: String,
+                               targetIdColumn: String)
+    extends WithStateId(
+      dataModel,
+      stateColumn,
+      orderColumn,
+      targetIdColumn, {
+        val window = Window.partitionBy(dataModel.idColumn).orderBy(orderColumn)
+        lag(stateColumn, 1)
+          .over(window)
+          .notEqual(col(stateColumn))
+          .or(lag(stateColumn, 1).over(window).isNull)
+      }
+    )
