+++ src/test/scala/io/arlas/data/transform/features/WithFragmentVisibilityFromTempoTest.scala
@@ -17,9 +17,10 @@
  * under the License.
  */
 
-package io.arlas.data.transform
+package io.arlas.data.transform.features
 
 import io.arlas.data.sql._
+import io.arlas.data.transform.ArlasTest
 import io.arlas.data.transform.ArlasTransformerColumns._
 import io.arlas.data.transform.VisibilityChange._
 import org.apache.spark.sql.types.{DoubleType, StringType, StructField}
