+++ src/test/scala/io/arlas/data/transform/timeseries/WithDurationFromIdTest.scala
@@ -17,12 +17,17 @@
  * under the License.
  */
 
-package io.arlas.data.transform
+package io.arlas.data.transform.timeseries
 
-import io.arlas.data.sql._
+import io.arlas.data.transform.{
+  ArlasTest,
+  ArlasVisibilityStates,
+  WithArlasVisibilityStateFromTimestamp
+}
 import io.arlas.data.transform.ArlasTransformerColumns._
 import org.apache.spark.sql.DataFrame
 import org.apache.spark.sql.functions._
+import io.arlas.data.sql._
 
 class WithDurationFromIdTest extends ArlasTest {
